+++ hutool-captcha/src/main/java/cn/hutool/captcha/GifCaptcha.java
@@ -5,7 +5,6 @@ import cn.hutool.captcha.generator.CodeGenerator;
 import cn.hutool.captcha.generator.RandomGenerator;
 import cn.hutool.core.img.ImgUtil;
 import cn.hutool.core.img.gif.AnimatedGifEncoder;
-import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.RandomUtil;
 
 import java.awt.AlphaComposite;