+++ hutool-poi/src/main/java/cn/hutool/poi/excel/sax/ExcelSaxUtil.java
@@ -5,6 +5,7 @@ import cn.hutool.core.date.DateUtil;
 import cn.hutool.core.exceptions.DependencyException;
 import cn.hutool.core.io.IORuntimeException;
 import cn.hutool.core.util.CharUtil;
+import cn.hutool.core.util.NumberUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.poi.excel.ExcelDateUtil;
 import cn.hutool.poi.excel.sax.handler.RowHandler;
@@ -264,7 +265,15 @@ public class ExcelSaxUtil {
 		if (StrUtil.isBlank(value)) {
 			return null;
 		}
-		return getNumberValue(Double.parseDouble(value), numFmtString);
+
+		// issue#IB0EJ9 可能精度丢失，对含有小数的value判断并转为BigDecimal
+		final double number = Double.parseDouble(value);
+		if(StrUtil.contains(value, CharUtil.DOT) && !value.equals(Double.toString(number))){
+			// 精度丢失
+			return NumberUtil.toBigDecimal(value);
+		}
+
+		return getNumberValue(number, numFmtString);
 	}
 
 	/**